database_script/requirements.txt
psycopg2-binary==2.9.9
pymongo==4.6.0
python-dotenv==1.0.0
requests==2.31.0
cloudinary==1.36.0
faker==20.1.0
